+++ EventFilter/CSCRawToDigi/plugins/CSCDigiValidator.cc
@@ -8,7 +8,7 @@
 #include <algorithm>
 #include <iostream>
 
-#include "EventFilter/CSCRawToDigi/interface/CSCDigiValidator.h"
+#include "EventFilter/CSCRawToDigi/plugins/CSCDigiValidator.h"
 
 #include "FWCore/Framework/interface/Event.h"
 #include "FWCore/ParameterSet/interface/ParameterSet.h"
